Clean up welcome and notification details compose views
