+++ app/src/main/java/io/homeassistant/companion/android/settings/notification/NotificationDetailFragment.kt
@@ -3,35 +3,18 @@ package io.homeassistant.companion.android.settings.notification
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.os.Parcelable
-import android.text.Html.fromHtml
 import android.view.LayoutInflater
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.viewinterop.AndroidView
 import androidx.fragment.app.Fragment
-import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.google.android.material.composethemeadapter.MdcTheme
 import io.homeassistant.companion.android.R
 import io.homeassistant.companion.android.database.AppDatabase
 import io.homeassistant.companion.android.database.notification.NotificationItem
-import kotlinx.android.parcel.Parcelize
-import java.util.Calendar
-import java.util.GregorianCalendar
+import io.homeassistant.companion.android.settings.notification.views.LoadNotification
 
 class NotificationDetailFragment(
     private val notification: NotificationItem
@@ -80,69 +63,6 @@ class NotificationDetailFragment(
                     LoadNotification(notification)
                 }
             }
-        }
-    }
-
-    @Parcelize
-    data class NotificationData(val received: Long, val message: String, val data: String) : Parcelable
-
-    @Composable
-    private fun LoadNotification(notification: NotificationItem) {
-        val notificationItem = rememberSaveable {
-            NotificationData(notification.received, notification.message, notification.data)
-        }
-
-        Column {
-            Text(
-                text = getString(R.string.notification_received_at),
-                fontWeight = FontWeight.ExtraBold,
-                fontSize = 20.sp,
-                modifier = Modifier
-                    .padding(top = 30.dp, bottom = 20.dp, start = 10.dp)
-            )
-            val cal: Calendar = GregorianCalendar()
-            cal.timeInMillis = notificationItem.received
-            Text(
-                text = cal.time.toString(),
-                modifier = Modifier
-                    .padding(start = 20.dp)
-            )
-            Text(
-                text = getString(R.string.notification_message),
-                fontWeight = FontWeight.ExtraBold,
-                fontSize = 20.sp,
-                modifier = Modifier
-                    .padding(top = 20.dp, bottom = 20.dp, start = 10.dp)
-            )
-            AndroidView(factory = { context ->
-                TextView(context).apply {
-                    text = fromHtml(notificationItem.message)
-                    setPadding(80, 0, 0, 0)
-                    textSize = 16f
-                }
-            })
-            Text(
-                text = getString(R.string.notification_data),
-                fontWeight = FontWeight.ExtraBold,
-                fontSize = 20.sp,
-                modifier = Modifier
-                    .padding(top = 20.dp, bottom = 20.dp, start = 10.dp)
-            )
-            val notifData =
-                try {
-                    val mapper = ObjectMapper()
-                    mapper.configure(JsonParser.Feature.ALLOW_UNQUOTED_FIELD_NAMES, true)
-                    val jsonObject = mapper.readValue(notificationItem.data, Object::class.java)
-                    mapper.writerWithDefaultPrettyPrinter().writeValueAsString(jsonObject)
-                } catch (e: Exception) {
-                    notificationItem.data
-                }
-
-            Text(
-                text = notifData,
-                modifier = Modifier
-                    .padding(start = 20.dp)
-            )
         }
     }
