+++ app/src/main/java/io/homeassistant/companion/android/util/PreviewData.kt
@@ -1,0 +1,5 @@
+package io.homeassistant.companion.android.util
+
+import io.homeassistant.companion.android.database.notification.NotificationItem
+
+val notificationItem = NotificationItem(1, 1636389288682, "testing", "{\"message\":\"test\"}")
